Create Pro_Assi_4.c
Write a C program to find the sum of two 1D integer arrays ‘A’  and ‘B’ of same size and store the result in another array ‘C’, where the size of the array and the elements of the array are taken as input.
In the Test case the input is given as follows
5
10
20
30
40
50
1
2
3
4
5
So the output will be displayed as
Result is
11
22
33
44
55
Write the program accordingly. Use dynamic memory allocation.